+++ src/main/scala/wdlTools/syntax/v1_1/ParseAll.scala
@@ -69,15 +69,13 @@ case class ParseAll(followImports: Boolean = false,
           }, loc)
 
         // string place holders
-        case CST.ExprPlaceholderEqual(t, f, value, srcText) =>
-          AST.ExprPlaceholderCondition(translateExpr(t),
-                                       translateExpr(f),
-                                       translateExpr(value),
-                                       srcText)
-        case CST.ExprPlaceholderDefault(default, value, srcText) =>
-          AST.ExprPlaceholderDefault(translateExpr(default), translateExpr(value), srcText)
-        case CST.ExprPlaceholderSep(sep, value, srcText) =>
-          AST.ExprPlaceholderSep(translateExpr(sep), translateExpr(value), srcText)
+        case CST.ExprPlaceholder(t, f, sep, default, value, srcText) =>
+          AST.ExprPlaceholder(t.map(translateExpr),
+                              f.map(translateExpr),
+                              sep.map(translateExpr),
+                              default.map(translateExpr),
+                              translateExpr(value),
+                              srcText)
 
         // operators on one argument
         case CST.ExprUnaryPlus(value, loc) =>