+++ src/test/scala/wdlTools/syntax/v1/ConcreteSyntaxV1_1Test.scala
@@ -367,7 +367,12 @@ class ConcreteSyntaxV1_1Test extends AnyFlatSpec with Matchers {
       case ExprString("\n    echo ", _) =>
     }
     task.command.parts(1) should matchPattern {
-      case ExprPlaceholderSep(ExprString(",", _), ExprIdentifier("min_std_max_min", _), _) =>
+      case ExprPlaceholder(None,
+                           None,
+                           Some(ExprString(",", _)),
+                           None,
+                           ExprIdentifier("min_std_max_min", _),
+                           _) =>
     }
   }